Routing Setup
- Tested in dev and it works, need to test in production, but change to server.js should allow it to work in production.
