+++ backend/server.js
@@ -36,7 +36,7 @@ router.get('/some_route', (req, res) => {
 if (process.env.NODE_ENV === "production") {
   router.use(express.static(path.resolve(__dirname, "../prod-frontend")));
 
-  router.get('/', (req, res) => {
+  router.get('*', (req, res) => {
     res.sendFile(path.resolve(__dirname,"index.html"))
   });
