+++ dev-frontend/buildEnvConfigs/webpack.dev.js
@@ -5,6 +5,7 @@ module.exports = merge(common, {
     mode: 'development',
     devtool: 'inline-source-map',
     devServer: {
+        historyApiFallback: true,
         proxy: {
             '/some_route': 'http://localhost:5000'
         }
